try lower precision

+++ src/data/earthengine/era5.py
@@ -33,4 +33,5 @@ def get_single_era5_image(region: ee.Geometry, start_date: date, end_date: date)
         return create_placeholder(region, ERA5_BANDS).toDouble()
 
     # all imagery has to have the same data type to be compatible
-    return image.toDouble()
+    #return image.toDouble()
+    return image
