+++ src/data/earthengine/modis.py
@@ -37,4 +37,5 @@ def get_single_modis_image(region: ee.Geometry, start_date: date, end_date: date
         return create_placeholder(region, MODIS_BANDS).toDouble()
 
     # all imagery has to have the same data type to be compatible
-    return image.toDouble()
+    #return image.toDouble()
+    return image
