+++ src/data/earthengine/s2.py
@@ -61,4 +61,5 @@ def get_single_s2_image(region: ee.Geometry, start_date: date, end_date: date) -
         return create_placeholder(region, S2_BANDS).toDouble()
 
     # all imagery has to have the same data type to be compatible
-    return image.toDouble()
+    #return image.toDouble()
+    return image
